+++ app/controllers/jobs_controller.rb
@@ -25,11 +25,15 @@ class JobsController < ApplicationController
   end
 
   def edit
-    # implement on your own!
+    @company = Company.find(params[:company_id])
+    @job = Job.find(params[:id])
   end
 
   def update
-    # implement on your own!
+    @company = Company.find(params[:company_id])
+    @job = Job.find(params[:id])
+    @job.update(job_params)
+    redirect_to company_job_path(@company, @job)
   end
 
   def destroy
