+++ app/views/jobs/show.html.erb
@@ -12,4 +12,6 @@ City: <%= @job.city %>
 
 Company: <%= link_to @job.company.name, company_path(@job.company) %>
 
+<%= link_to 'Edit Job Record', edit_company_job_path(@job.company) %>
 <%= link_to "Delete Job Record", company_job_path(@job.company), method: :delete %>
+
